lib: extract deprecation into separate file
