+++ lib/mime_actor.rb
@@ -6,30 +6,13 @@
 require "mime_actor/version"
 require "mime_actor/errors"
 
-require "active_support/dependencies/autoload"
-require "active_support/deprecation"
+# load mime_actor components
+require "mime_actor/action"
+require "mime_actor/callbacks"
+require "mime_actor/logging"
+require "mime_actor/rescue"
+require "mime_actor/scene"
+require "mime_actor/stage"
+require "mime_actor/validator"
 
-module MimeActor
-  extend ActiveSupport::Autoload
-
-  autoload :Action
-  autoload :Scene
-  autoload :Stage
-  autoload :Rescue
-  autoload :Validator
-  autoload :Logging
-
-  def self.deprecator
-    @deprecator ||= ActiveSupport::Deprecation.new("0.7.0", "MimeActor")
-  end
-
-  [
-    [MimeActor::Rescue::ClassMethods, { rescue_actor: "use #rescue_actor instance method" }],
-    [MimeActor::Scene::ClassMethods, { act_on_format: :respond_act_to }],
-    [MimeActor::Stage::ClassMethods, { actor?: "no longer supported, use Object#respond_to?" }],
-    [MimeActor::Stage::ClassMethods, { dispatch_cue: "no longer support anonymous proc with rescue" }],
-    [MimeActor::Stage::ClassMethods, { dispatch_act: "no longer support anonymous proc with rescue" }]
-  ].each do |klazz, *args|
-    deprecator.deprecate_methods(klazz, *args)
-  end
-end
+require "mime_actor/deprecator"
